Added some more rake tasks related to installation/uninstallation.  Added confirmation questions for destructive actions.

+++ lib/tasks/ratings_extension_tasks.rake
@@ -1,6 +1,21 @@
 namespace :radiant do
   namespace :extensions do
     namespace :ratings do
+      desc "Single task to install and update the ratings extension"
+      task :install => [:environment, :migrate, 'update:all']
+
+      desc "Single task to uninstall the ratings extension"
+      task :uninstall => :environment do
+        if confirm('Are you sure you want to uninstall the ratings extension?  This will revert the migration, remove the ratings assets, and delete the rating snippets and config settings.')
+          ENV['FORCE'] = 'true'
+          %w(revert_migration delete:all).each { |t| Rake::Task["radiant:extensions:ratings:#{t}"].invoke }
+        end
+      end
+
+      def confirm(question)
+        require 'highline/import' unless respond_to?(:agree)
+        ENV['FORCE'] || agree("#{question} [yn]")
+      end
 
       desc "Runs the migration of the Ratings extension"
       task :migrate => :environment do
@@ -9,6 +24,14 @@ namespace :radiant do
           RatingsExtension.migrator.migrate(ENV["VERSION"].to_i)
         else
           RatingsExtension.migrator.migrate
+        end
+      end
+
+      desc "Reverts the migrations run by the ratings extension"
+      task :revert_migration => :environment do
+        if Rating.table_exists? && confirm("This task will destroy all your ratings data. Are you sure you want to continue?")
+          require 'radiant/extension_migrator'
+          RatingsExtension.migrator.migrate(0)
         end
       end
 
@@ -28,9 +51,11 @@ namespace :radiant do
             puts "Updating config settings..."
             get_yaml('config_defaults.yml').each do |key, value|
               c = Radiant::Config.find_or_initialize_by_key("ratings.#{key}")
-              c.value = value
-              c.save!
-              puts "  - ratings.#{key} = #{value}"
+              if c.new_record? || confirm("Are you sure you want to overwrite existing config setting ratings.#{key}?")
+                c.value = value
+                c.save!
+                puts "  - ratings.#{key} = #{value}"
+              end
             end
           else
             puts "The radiant config table does not exist.  Please create it, then re-run this migration."
@@ -45,9 +70,11 @@ namespace :radiant do
 
             path = file.sub(RatingsExtension.root, '')
             directory = File.dirname(path)
-            mkdir_p RAILS_ROOT + directory, :verbose => false
-            cp file, RAILS_ROOT + path, :verbose => false
-            puts "  - #{path}"
+            if !File.exists?(RAILS_ROOT + path) || confirm("Are you sure you want to overwrite existing asset #{path}?")
+              mkdir_p RAILS_ROOT + directory, :verbose => false
+              cp file, RAILS_ROOT + path, :verbose => false
+              puts "  - #{path}"
+            end
           end
         end
 
@@ -56,9 +83,11 @@ namespace :radiant do
           puts "Updating snippets for the ratings extension..."
           get_yaml('snippets.yml').each do |snippet_name, snippet_content|
             s = Snippet.find_or_initialize_by_name(snippet_name)
-            s.content = snippet_content
-            s.save!
-            puts "  - #{snippet_name}"
+            if s.new_record? || confirm("Are you sure you want to overwrite existing snippet #{snippet_name}?")
+              s.content = snippet_content
+              s.save!
+              puts "  - #{snippet_name}"
+            end
           end
         end
       end
@@ -72,7 +101,7 @@ namespace :radiant do
           if Radiant::Config.table_exists?
             puts "Deleting config settings..."
             get_yaml('config_defaults.yml').each do |key, value|
-              if c = Radiant::Config.find_by_key("ratings.#{key}")
+              if c = Radiant::Config.find_by_key("ratings.#{key}") and confirm("Are you sure you want to delete config setting ratings.#{key}?")
                 c.destroy
                 puts "  - ratings.#{key} = #{value}"
               end
@@ -86,7 +115,7 @@ namespace :radiant do
           Dir[RatingsExtension.root + "/public/**/*"].each do |file|
             next if File.directory?(file)
             path = file.sub(RatingsExtension.root, '')
-            if File.exists?(RAILS_ROOT + path)
+            if File.exists?(RAILS_ROOT + path) && confirm("Are you sure you want to delete asset #{path}?")
               rm RAILS_ROOT + path, :verbose => false
               puts "  - #{path}"
             end
@@ -97,7 +126,7 @@ namespace :radiant do
         task :snippets => :environment do
           puts "Deleting snippets installed by the ratings extension..."
           get_yaml('snippets.yml').each do |snippet_name, snippet_content|
-            if s = Snippet.find_by_name(snippet_name)
+            if s = Snippet.find_by_name(snippet_name) and confirm("Are you sure you want to delete snippet #{snippet_name}?")
               s.destroy
               puts "  - #{snippet_name}"
             end
